+++ src/types/index.ts
@@ -1,5 +1,5 @@
 export * from "./payload";
 export * from "./label";
 export * from "./handlers";
-export * from "./configuration";
+export * from "./configuration-types";
 export * from "./markdown";
